refactors module to config
